+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/Body.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum Body {
+    SKINNY, SLIM, MUSCULAR, CHUBBY
+}
