+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/Channel.java
@@ -1,4 +1,4 @@
-package com.yapp.lonessum.domain.meeting.entity.constant;
+package com.yapp.lonessum.domain.constant;
 
 public enum Channel {
     FACEBOOK, INSTAGRAM, KAKAOROOM, KAKAOPLUS, FRIEND, COMMUNITY
