+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/Character.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum Character {
+    VERY_QUIET, A_LITTLE_QUIET, VERY_ACTIVE, A_LITTLE_ACTIVE
+}
