+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/DateCount.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum DateCount {
+    ZERO, ONETWO, THREEFOUR, FIVE
+}
