+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/DomesticArea.java
@@ -1,4 +1,4 @@
-package com.yapp.lonessum.domain.meeting.entity.constant;
+package com.yapp.lonessum.domain.constant;
 
 public enum DomesticArea {
     ICN, SNW, SNE, SSW, SSE, GN, GS
