+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/Gender.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum Gender {
+    MALE, FEMALE
+}
