+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/Mindset.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum Mindset {
+    ALL, FRIEND, LOVE
+}
