+++ lonessum/src/main/java/com/yapp/lonessum/domain/constant/TypeOfMeeting.java
@@ -1,0 +1,5 @@
+package com.yapp.lonessum.domain.constant;
+
+public enum TypeOfMeeting {
+    ONE, TWO, THREE, FOUR
+}
