--- lonessum/src/main/java/com/yapp/lonessum/domain/meeting/entity/constant/Gender.java
@@ -1,5 +1,0 @@
-package com.yapp.lonessum.domain.meeting.entity.constant;
-
-public enum Gender {
-    MALE, FEMALE
-}
