--- lonessum/src/main/java/com/yapp/lonessum/domain/meeting/entity/constant/Mindset.java
@@ -1,5 +1,0 @@
-package com.yapp.lonessum.domain.meeting.entity.constant;
-
-public enum Mindset {
-    ALL, FRIEND, LOVE
-}
