--- lonessum/src/main/java/com/yapp/lonessum/domain/meeting/entity/constant/TypeOfMeeting.java
@@ -1,5 +1,0 @@
-package com.yapp.lonessum.domain.meeting.entity.constant;
-
-public enum TypeOfMeeting {
-    ONE, TWO, THREE, FOUR
-}
